Added flee chance and pop-up screens.
Added percentage chance that attempting to flee from battle will fail, causing the enemy to attack for free that turn (this chance is based on speeds of both parties). Also added both a pause menu which allows closing of the game as well as checking the party's stats, and a post-battle screen which displays when winning a battle.