Added battle waits, and cameras
Altered the battle attack calls to use delays that last for a few seconds so each individual attack is seen (this is combined with a label that displays the name of the attack). Also added multiple different cameras for each player party member, which switches depending on which ones action is being selected (swapping to the main camera for the actual attacks).