Moved turn attack functions to C++
Removed a large amount of the main blueprint for controlling the launching of attacks during battle. Have replaced this with a much more compact and efficient section of code. Also replaced the random target function blueprint with C++, and revised the selection of enemy attacks to be done in C++ too.