Added magic stat.
Added a magic stat to combatants that is used to factor the values for magic attacks and healing skills. Attack functions have been updated to account for this, and attacks in the data tables have now been assigned "types" to determine if they are physical, magical, or healing. This can be expanded on in the future to account for other types (buffs, etc.).